+++ packages/reactivity-core/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.4.3 (Unreleased)
 
 -   Introduce `subtleWatchDirty`, a function that allows one to watch for signal changes without triggering the re-evaluation of the signal.
+-   Add missing `forEach()` method to `ReactiveSet` and `ReactiveMap`.
 -   Deprecate `syncEffectOnce` (use `subtleWatchDirty` instead).
 
 ## v0.4.2